Make pyarts 2.5.10 work as line frontend
The units are hopefully correct.  ARTS produces output in 1/m but pyLBL wants
it in 1/m/mol, as far as I understand it.

Note also that this only takes the absorption data as presented by pyLBL and converts
it into the ARTS format.  It makes no decisions about continua (which if present should
set the cutoff of water-lines to 750 GHz), and there is no data about line-mixing
available at this stage, so O2, CO2, and CH4 will produce poor simulation results